+++ examples/wasm-example/src/lib.rs
@@ -1,4 +1,4 @@
-use libipld::ipld;
+use ipld_core::ipld;
 use rust_ipfs::UninitializedIpfsNoop as UninitializedIpfs;
 use rust_ipfs::{Multiaddr, Protocol};
 use wasm_bindgen::prelude::*;