+++ unixfs/src/symlink.rs
@@ -33,8 +33,8 @@ pub fn serialize_symlink_block(target_path: &str, block_buffer: &mut Vec<u8>) {
 mod tests {
     use super::serialize_symlink_block;
     use core::convert::TryFrom;
-    use libipld::multihash::{self, Multihash};
-    use libipld::Cid;
+    use ipld_core::cid::Cid;
+    use multihash::{self, Multihash};
     use sha2::{Digest, Sha256};
 
     #[test]
@@ -46,7 +46,11 @@ mod tests {
         // `foo_directory/b`.
         serialize_symlink_block("b", &mut buf);
 
-        let mh = Multihash::wrap(multihash::Code::Sha2_256.into(), &Sha256::digest(&buf)).unwrap();
+        let mh = Multihash::wrap(
+            multihash_codetable::Code::Sha2_256.into(),
+            &Sha256::digest(&buf),
+        )
+        .unwrap();
         let cid = Cid::new_v0(mh).expect("sha2_256 is the correct multihash for cidv0");
 
         assert_eq!(